+++ NEWS.md
@@ -1,3 +1,24 @@
+# CHANGES IN nestr VERSION 0.2
+
+## NEW FEATURES
+
+## BUG FIXES
+
+## MAJOR CHANGES
+
+- package now uses `typr` for implicitly typed objects when `reactive = FALSE`
+- modified: default value of `gap` --> `TRUE` (in `setNested()`)
+
+## MINOR CHANGES
+
+- modified: `README` (some typos)
+- modified: unit tests (`test-setNested.r`)
+- modified: examples (`setNested.r`)
+
+## MISC
+
+-----
+
 # CHANGES IN nestr VERSION 0.1
 
 ## NEW FEATURES